conversational-assistant/requirements.txt
requests
azure-ai-textanalytics
azure-identity
python-dotenv